fix : modify saint model

+++ code/sequence_/sequence_utils/models.py
@@ -243,7 +243,7 @@ class Saint(nn.Module):
         self.enc_dec_mask = None
 
     def get_mask(self, seq_len):
-        mask = torch.from_numpy(np.triu(np.ones((seq_len, seq_len)), k=1))
+        mask = torch.from_numpy(np.triu(np.ones((seq_len, seq_len)), k=1)).float()
 
         return mask.masked_fill(mask == 1, float("-inf"))
